+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## Current
 ### Features
 ### Fix
+* Add missing create button in Phone index
+* Add missing create button in POI index
+* Add missing create button in users index
 * Reservation scope can be set in days on base creation
 ## Version 1.4.5
 ### Features